Resource definitions
Hop on the bandwagon! Everyone and their brother implements Copper Ore,
why not us?  Malachite's here, too.  Full block recipes (to and from),
as well as smelting for Copper Ore to ingots.

Partial fix for #9 - still have to work out the incense plant, and none
of this is yet available by any means other than the Creative tab.

Also, refactored HekaCore so that the actual load() method is a bit
easier to read (and moved everything that was in preInit() down to
load() - can shuffle things around again if need be, but for now it's
not required)