Incense Plant
Based on myrrh, in terms of graphics and such. The item doesn't do
anything yet, but it's there and spawning in the world (only in
deserts). Harvesting with a pesheskef (right-click) will grant resin
without mangling the plant into unusability - destroying with either the
pesheskef or shears will produce resin, but kill the plant. Using
improper tools will get you NOTHING.