Stelae - now with signs!
Hooked into the same GUI and NBT structure as vanilla signs use, so they
have the same limitations. In order to use those hooks, sadly, needed a
code rework to have separate Item classes for the blocks (instead of
just coasting on ItemBlock), which means the Stonepainter's Workbench is
now a little bit messier to deal with not all of its output being
Blocks.

Partial solution to #7 - keeping it open for additional effects to be
added in the future.